Refactor classes into individual files and update imports
